putting detection & thresholds in library
MVP working camera file right now is comm_stop_for_orange.py. detection.py and thresholds.py are libraries that the comm file uses (libraries already downloaded on nicla, comm is main.py on nicla), which I made to compartamentalize the code so the main.py is mostly just doing the work of getting the image, getting a message based on the image (which is interpreted based on the chosen function from detection.py i.e. color blobs or distance or line tracking), and sending that message over USB serial to the raspi brain